+++ tests/state/test_schemes_api.py
@@ -10,7 +10,7 @@ class TestSchemeAPI:
             "offset": 0,
             "selected_actors": ["actor"],
             "selected_statuses": ["Success"],
-            "selected_ids": ["id"],
+            "selected_message_ids": ["id"],
             "start_datetime": "2021-08-16 10:00:00",
             "end_datetime": "2021-08-16 10:00:00",
         }
@@ -25,7 +25,7 @@ class TestSchemeAPI:
             "offset": "offset",
             "selected_actors": [1],
             "selected_statuses": ["status"],
-            "selected_ids": [123],
+            "selected_message_ids": [123],
             "start_datetime": 100000,
             "end_datetime": 100000,
         }
@@ -34,7 +34,7 @@ class TestSchemeAPI:
             "end_datetime": ["Not a valid datetime."],
             "offset": ["Not a valid integer."],
             "selected_actors": {0: ["Not a valid string."]},
-            "selected_ids": {0: ["Not a valid string."]},
+            "selected_message_ids": {0: ["Not a valid string."]},
             "selected_statuses": {0: ["Must be one of: Started, Pending, Skipped, Canceled, Failure, Success."]},
             "size": ["Must be greater than or equal to 1 and less than or equal to 1000."],
             "sort_column": ["Not a valid string."],